move purge functionality to purge.py
